services/github_readme_scraper/requirements.txt
PyGithub>=1.55
python-dotenv>=0.19.2
